pal: GKV customization with usecasemanager xml
Retrieve GKVs from usecasemanager xml file in
payload builder instead of hardcoding the same.

Change-Id: I1f360aafc6ffbd1ea4870efa3c5d385407fde117

pal: Add fallback mechanism for custom config as selector

When a custom config is set and it doesn't need GKV pair
different than the default one, current implementation requires
it to be added in usecase manager XML to pick the GKVs.
Add a fallback mechanism to find the KVs by removing the
custom config as selector when the KV pairs are not
found with it.
Modify the way to get selector values for custom config using
device attributes.

Change-Id: I66c11b1ac92c04f65b3e6ee57e5b838f95f7402d

pal: Update the device derivation logic in device and devicepp KV

In order to populate device and devicepp KVs, the associated devices
used to be obtained from stream. Instead derive the device using
getInstance from device class to avoid the stream dependency as
in some cases like external ec reference devices, stream can be NULL.

Change-Id: Ibc2a7a26ba3b078f6de8808ac2ac58d5236e1361

pal: Sort the key value tags in usecase manager XML

When there are no selector values or filled selector pairs,
the default KV pair which has no selectors should be picked.
But it currently picks the first tag KV pairs. Sort the key value
tags as per number of selectors to fix this.

Change-Id: I1a1a9e7992abc0484597d381656208ce2746f1aa

pal: Clean up the KV pair handling

Remove the KV pair handling logic in resource
manager as it is moved to usecase manager in
payload builder.

Change-Id: Id38c074cc3675f63314cbbcdaf53256078ee5162

temp: fix compilation for gkv

Change-Id: I8a0264acc92c6767613a2997da518bd0507d0375