Pal : Deprecate stream mute and unmute on USB removal.
For voice call, common Stream is there for RX and TX,
stream mute is set and unset on TX when mute setConfig
is called on USB removal. so client side still receives
audio from DUT.

Deprecate stream mute and unmute set config as part of force
device switch(SPKR->SPKR) to recover FET. Spkr boost enable
is a blocking call to PMIC driver, this ensures enough 
time for FET to recover so mute/unmute is not needed.

Change-Id: Iec1196f843d21eb7d61ac7ad86ee004100c1ea17