pal: Move concurrent stream handling to open and close
Notify concurrent stream in open and close instead of stream start
and stop as the memory usage of graph modules in SPF peaks
during LPI->NLPI or vice versa if it is done in start or stop
sequence. As switch sequence is moved to open, remove stream active
check while getting concurrency count.

Change-Id: I4aea9933ccab76ad97344607020b4d1726c455f7