pal: fix HFP uplink issue with GAS enable
On Android S, handset_mic is default opened after LA up if GAS stream enabled.
Speaker_mic is working as HFP's uplink input and it uses the same TDM
interface config with handset_mic.
In previous handling, Pal will try to switch device for the incoming HFP case,
but it failed because GAS is always working.
Hence with this change, we prevent streamDevSwitch when the incoming stream is
HFP type.

Change-Id: Ic5777d4d9683c16da0a802910ec8df3d037bceff