pal: move compress_start to stream start
Earlier, in order to calculate timestamp for audio buffer,
we wanted to start stream in ADSP as close to first read.
Hence compress_start is in read call. However, the decision
and calculation of timestamps is left to Audio HAL.
Now, move the compress_start to session start so that it executes
with AudioPerf lock accquired. This change contributes to lessen
the coldstart latency of the audio recording

Change-Id: I1902d85903d09580655173e8465d72aa68702b63