pal: remove PSPD MFC configuration
This change extends commit 12b3003ede98d02b130ad5a0ef2ae3797431bbb5
to Linux based platform that uses PAL.

For automotive use cases that use TDM, the PSPD MFC is configured by PAL
using the TDM group device configuration, which results in the output of
the PSPD MFC to be 8 channels for media playback use case. This will
increase the processing required in the DSP for the device-pp sub-graph.

Removed PSPD MFC and disabled the code that configures it.