PAL: Handle subtraction overflow
While calculating bytes_written, if position_frames value is less than
mmap_write_position_, we will get negative value, and we pass it to
FrameToBytes, but it takes only unsigned value for its argument, hence
we will get this subtaction overflow.

Now, we will first check if resultant bytes_written is negative or not,
if its negative we return with error.

Change-Id: I43fb5954e7517f360b9056760d58fd5dea19eaff