pal: avoid implicit device switch during resume if call is active
If call is active on SCO and in-between SCO mode gets disconnected,
AudioDeviceBroker will send resume params for A2DP profile. This will
cause erroneous routing to A2DP devices during the call.

Added param to check if phone mode is not normal and avoid device
switch in that case.

Change-Id: I628890aed0e233e6ce76ca831aac3ff5f86d60d8