ar-pal: acquire stream lock in drain()
This is to fix potential deadlock issue:
thread-1: calls into flush() with agm session lock
          held, and in it's callback function, it
          can wait for the early_eos_lock in plugin.
thread-2: calls into drain() which can hold the
          early_eos_lock in plugin and then wait on
          agm session lock.
With the above sequence, it can lead to deadlock.
Fix this by acquiring stream lock in drain() to
synchronize with flush() which can also acquire
stream lock.

Change-Id: I1ef2ae69dee4f7a9942b5e443c11b871db304db7