pal: do not reconfigure SCO sample rate if SCO is not connected
If BT device switch happens during active call, first SCO gets
disconnected for the old device and sco prams are received for
the new device. If there is a param mismatch and there are active
streams over SCO due to race condition, doing force reconfigure
will lead to graph errors as SCO is not connected yet for the new
device.

Fixed this by only doing force reconfigure if SCO is connected and
there is a mismatch is SCO params with active streams.

Change-Id: I6efab617d2b2b85398988a4f0e2d141db6581e6a