pal: ignore EALREADY failure in sound trigger usecase prepare
In prepare of sound trigger usecase, we will explicitly set ckv
for device/device pp. When SVA and ACD are resumed with different
device configuration, ACD may fail in calibration set as it is
already set in SVA resume and hw ep is already started, and this
failure will cause ACD to fail in start.

Ignore EALREADY failure for ckv update in sound trigger usecase
prepare so that sound trigger usecase can continue with start
operation.

Change-Id: I0a133854325ea037a57c0ce527a01026ce083c60