Pal: Fix for session callback registration failures
Any feature that was using a session callback is failing.
This issue was due to a race condition where the resource manager
attempted to use the callback before it was properly registered.
This fix moved the callback registration earlier in the open process
to avoid this race condition.
This changed also fixed a failure in setting the initial volume
by fixing a improper tag value.

Change-Id: I1eb320b6ab884ebb73f87b7a5a4f8bc6f73d51cc