audio pal: Correct Custom Param API
(1)Correct old GEF_PARAM api because it is removed
for pal_gef_rw_param. Add some log to check;

(2)The payload param should be transfer to server for pal_get_custom_param because aidlret data can't get in the
ipc server. and it need too use payload tagid for pal.

Change-Id: Ifd4bc82370bdacf64801d031f416b724397e7620