PAL: Add timeout event support
For timeout cases, where there is no speech for few seconds, for e.g,
5 seconds for non-continuous mode, algo will send timeout event, which
which need to propagate to apk, apk can take appropriate action to
handle it based on its expectation, one possible way is stop the
session, and then restart it.

Previously, it was treated as failure case and did not propagated
to apk, now we need to propagate it to apk.

Change-Id: I981dd5222c6f07b1571579d6c6255c28618585b9
(cherry picked from commit 06668d00edfcb7229bc847eb8eec5ba9e05fe2c6)