pal: clear snd device name to avoid unexpected override
If stale snd device name for overriding is not cleared
when there is no active use case on this device, then
it may override snd device name for the new use case,
leading to wrong device path applied.
Clear the stale snd device name if there is no active use
case on this device.

Change-Id: I313c67e71d3a8234db4b2d3bd5723feb5ff43022