PAL: Refactor XML parsing logic for sound trigger usecases
For parsing a tag, while loop was used but it was not needed,
as at once only one param under one tag was sent to get parsed,
and this while loop was adding complexity and an issue where
vendor_uuid's value was getting treated as param name, and while
parsing this invalid param's value, audioserver was getting crashed.

Now, removed the while loop while processing any tag.
And split the functions by adding getters for given attributes,
to simplify the tag handling function.

Change-Id: I314440bb8f370bcac664357e556ca80190ed9528