Bluetooth: add support for tuned latency for BT format
Although BT Host is responsible for sink latency,
which should be sufficient for entire BT latency
"to ear" latency.

Due to unaccounted latency, we may or need to tune
the latency based the codec chosen.

This adjustment must not be used to adjust any latency
ideally. This should carefully adjust only if required.

Change-Id: I202ba02873a50aa6e6ba9b0e9b5594c9f3d152da