PAL: Push abort event from event thread to apk in case of SSR
Previously, abort event was sent via same thread, in case of
SSR down state's force output setParam, this can trigger
deadlock as apk is expecting return status of setParam, but
PAL sends a callback event and waits for event processing at
apk side.

Now, updating the force output event handling in SSR down state,
to send a setParam to engine and push a abort event in queue and
process it in same way as other events.

Change-Id: I6788477973060d6e293b14f9c2e495956bbd6020