pal: rm: Update the snd card name in the init_audio
During bootup, audio_init will look for the soundcard
registered and enable the audio_mixer. In a corner case
if a SAMSUNG USB headset is connected during bootup
the audio_init is fetching the samsung headset sound card
instead of the actual sun-snd-card from codec.

This is due to the "SUN" substring is part of name Samsung.
Hence added '-' in the platform search string to avoid
picking the samsung headset as default sound card.

Change-Id: I21d98614ec605e1b4ec2e7bfefd498ec9c50d361