ar-pal: remove stream lock in drain()
This is to fix potential deadlock issue:
thread-1: calls into offloadThreadLoop() with compress obj lock
          held, and in the drain(), it
          can wait for the stream lock.
thread-2: calls into close() which can hold the
          stream lock and then wait on
          compress obj lock.
With the above sequence, it can lead to deadlock.
Fix this by removing stream lock in drain() .
Note: Current change is a revert change of 5333009.
The change 5333009 is merged for fix the issue CR3838229.
However, it can cause current deadlock issue and the CR3838229
can fixed by change 5390080. Therefore we revert it.

Change-Id: I2ec37ca8e4dd381e75b27649dc9de68f39fc702c