PAL: StreamHaptics: Allowing touch haptics when ringtone is active
When ringtone haptics ends there is a wait time of two seconds from
audioflinger.
In this two seconds there is a possibility that touch haptics can be
triggered from vibrator hal.
Now as per priority logic ringtone haptics is having more priority
which lets touch haptics to return failure because stream is active.
Removing this priority check after considering that the touch haptics
will only be received in PAL if ringtone is stopped from telepony and
the stream is in waiting state.
If in case touch haptics is received when ringtone is running and not
in waiting state, it is expected for ringtone vibration to stop touch
to continue.

Change-Id: I6e5bbb90f3ee8c40847307dc217222d0628da961