PAL: HapticDev: update haptics calibration mode logic
Current implementation allows processing mode to take over priority if
calibration mode is running and it will tear down calibration thread.
But if calibration mode is coming when processing mode is running there
is not proper syncronization which is leading to disrupt the haptics
graph state.
Adding a check to allow calibration mode only if haptics device is not
in use.

Change-Id: I40a05660e17a378c1051cd520e4eae8a23049926