PAL: Update the check for first EC set per stream
Previously, 'if' condition was checking if current rx device is
same as the previous one, then it was setting the flag for 'first
EC set' as true, which should not be the case.

Now update it to check if previous rx device is nullptr than, setting
'first EC set' as true'.

Change-Id: I353f3681514743dcedaecaacb456cb89a5ebc067