pal: stream: avoid race condition during deviceswitch
During device swtich there is chance that a muted shared stream
may be closed parallely which may lead to crash while unmuting the
stream post device switch.

Inorder to avoid this race condition, added these cached stream
to the RM usercounter so that the stream is not closed until released.

Change-Id: I87ea9e83cc707c6a8ebbc1324790619630be7625