PAL: Handle more than one resource available callback
Only one function pointer was cached for resource available callback,
as only STHAL was setting it, now QASR is also setting it, which
caused the overwriting of same function pointer, due to which older
pointer is lost and never triggered.

Now, caching all the resource available function pointers and the
cookies in respective vector.

Change-Id: I880f255388cc84ac97989b5cf11b7166510d2009