PAL: StreamHaptics: Update logic for handling stop wave design param
When pcmv haptics is running and ringtone usecase get trigger, it takes
priority and start running. Vibrator hal does not have the information
that ringtone haptics has taken over the usecase and currently it is
running in adsp, so when vibrator hal tries to stop the usecase it is
eventually stopping ringtone usecase.
Adding a logic to maintain current haptics running stream so that we can
use that stream to decide when we need to fire stop param, when ringtone
is active and stop is triggered by pcmv we need to skip stop.

Change-Id: I1f69c0f892ef70458a8f5d7243151c68e8ac0f49