PAL: HapticsDev: Update logic for using pcmRX id
when calibration is running and user triggered processing mode, pcmRxID
is allocated as part of stream setup in processing mode and then
calibration teardown happens in which we remove the frontends.
This is leading to session mode of RX set to default state which is
leading to issue when it is trying get miid for haptics gen module
present in RX subgraph and leading to pal_stream_start failure.
To handle such scenario pcmRX id allocated in calibration will be freed
in processing mode.

Change-Id: I454495e89a5bffe37c5da10c9bffa18c493c226f