pal: avoid wakelock release for device switch during detection
Wakelock release happens when device switch comes in detected
or buffering state, which will be a problem as device switch
does not change stream state. When client app tries to restart
for next detection, wake lock will be released again which is
unexpected.

Remove wakelock release for device switch in detected and
buffering state and wait for client restart/stop to trigger
wakelock release.

Change-Id: I027fd7bfad850779c55bff9af56091cc9a11c251