PAL: HapticsDev: update vote/unvote sleepmonitor logic for all mode
when haptics calibration fails for regular bootup calibration there is
a retry mechanism where calibration is retriggered, as unvoting is done
only in the exit of routine function of calibration thread this creates
a problem of unvoting happening lesser time in case of failure.
Adding unvote at the exit of HapticsDevStartCalibration to make sure it
is always unvoted.
For processing mode add a voting in device start and do unvote once the
caller returns as voting is happening in sessionalsapcm start.

Change-Id: I3decbb7cbe463c8ad5b2ab9e106f78f606cab6e5
(cherry picked from commit a898f714c2fb000d0f422cd984480f60e488c8f2)