PAL: Update check to accommodate paused ST stream
If any tx concurrency comes in when VUI stream is buffering switch
was deferred(in case of VUI stream LAB buffering), and rx concurrency
also got started, which was not supported, which cause VUI stream to
pause, and then tx stream ended but it can not update the defer flag
as VUI is not in buffering state and then after rx concurrency ended
VUI stream try to handle deferred switch, and try to start VUI stream
in NLPI, which cause a sleep monitor state mismatch, and results in
crash.

Now, updating the check, which will allow outgoing stream to update
deferred switch flag to 'NO_DEFER', without depending stream state,
as because of this stream VUI needed to switch.

Change-Id: I2819265016821f5a7502eb8619b582551ef95846