ar-pal: Update return value for inactive stream.
When a pluggable device is removed and set to NONE, an
error status may be returned if the stream is inactive,
preventing proper resource cleanup and affecting
subsequent reinsertion of the device.
pal_stream_set_device returns a success status when the
stream is inactive and the device is NONE to avoid it.

Change-Id: I7c3507a734aa8846b198cfd1cbddf294ddc46619