PAL: Add check for current charging state before voting for sleep monitor
When transit to NLPI on charging is enabled, and SVA starts in LPI and
vote for LPI to sleep monitor, and then playback stream starts, and
vote for NLPI, and it will cause SVA to switch to NLPI, but still vote
for LPI as vote is decided based on stream type, then charger is
connected, SVA is already in NLPI, so there will be no switch here. Then
playback stops, it will unvote NLPI for itself. Here, SVA will run in
NLPI, but sleep monitor will have LPI vote only, which will cause stat
mismatch in sleep monitor.

Now, updated the voting mechanism here, if charger is already connected
and then stream starts, irrespective of stream type, it will vote NLPI
for charger as well. And if charger got connected when at least one
stream is active(again irrespective of stream type), one NLPI vote for
charger will be done. During unvote, if charger vote was done, then the
last stream which will come to unvote itself, will unvote the charger's vote as well.

Change-Id: I381e9aa43fccfba7d52e1bdb2360bc40739ce420