PAL: Refine sleep monitor voting mechanism
Previously, voting was done based on static decision, if one stream
is classified as low power it will vote for low power, irrespective
of usecase state.

Now, updated the logic, to use current usecase state to vote, if
usecase is running in low power then it will vote for low power,
otherwise it will vote for high power.

Change-Id: I11b18b1ada731db13a54ecc26fa7072e956d675f