pal: Set Reader to Prepared when first stage is detected
In the current logic Reader is set to prepared state
when the detection is notified to the client.
So upon receiving a switch while a 2nd stage detection
is being processed it leads to failure since the switch
is handled rather being deferred.

Hence it is reverted back where the Reader now is set to
Prepared state when 1st stage is detected thus deferring
any incoming switch during detection.

Change-Id: I081fb592994d79c86dc93f4682b1b7965dd758c4
(cherry picked from commit 0efa12e3386553f2a685580949842b2519335d35)