Pal: fix double free issue in Session Alsa Get params.
The getParameters function takes the double pointer payload and can make it's copy into payloadData.

In case when the exit label is called in this function the payloadData get's freed and set to NULL but the orignal double pointer is not set to NULL thus resulting in double free when the call returns to the caller of getParameters.

Change-Id: I7958c1ffa7c4171cd2930a772d34df9d8d55dfde
Signed-off-by: ayush jain <ayjai@qti.qualcomm.com>
(cherry picked from commit bd9f6a6716da2d3668a4ebafa38e4fd72dbb396d)